third_party/blink/web_tests/fast/hidpi/image-srcset-ignore-double-descriptor-expected.txt

CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
CONSOLE WARNING: Failed parsing 'srcset' attribute value since it has multiple 'x' descriptors or a mix of 'x' and 'w'/'h' descriptors.
CONSOLE WARNING: Dropped srcset candidate "resources/blue-100-px-square.png"
PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("foo").clientWidth==100 is false
PASS document.getElementById("foo").clientWidth==200 is true
This test passes if the image below says 2x with a greenish background even when the deviceScaleFactor is 1.
